Designed cart layout i.e displaying all item carts and also add food quantities and on calculate there price

+++ app/src/main/java/in/ecommerce/takeaway/Database/LocalCartDataSource.kt
@@ -13,7 +13,7 @@ class LocalCartDataSource(private val carDAO:CartDAO):CartDataSource {
         return carDAO.countItemInCart(uid)
     }
 
-    override fun sumPrice(uid: String): Single<Long> {
+    override fun sumPrice(uid: String): Single<Double> {
         return carDAO.sumPrice(uid)
     }
